@@ -1,16 +1,15 @@
 name: masked_controller
 description: A controller to insert mask to textfield.
 version: 1.0.1
-author: Isaías Santana <isds.santana@gmail.com>
 homepage: https://github.com/IsaiasSantana/masked_controller
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.1.7
+  meta: ^1.8.0
 
 dev_dependencies:
   flutter_test:
