@@ -18,7 +18,10 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({
+    required this.title,
+    Key? key,
+  }) : super(key: key);
   final String title;
 
   @override
@@ -26,8 +29,7 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  final MaskedController _controller =
-      MaskedController(mask: Mask(mask: 'NNN.NNN.NNN-NN'));
+  final MaskedController _controller = MaskedController(mask: Mask(mask: 'NNN.NNN.NNN-NN'));
 
   @override
   void initState() {
